Get the bundle applier under the line ceiling it has been over for a while
`applier.py` was 762 lines against a 600 ceiling, and every change to it had to
buy its own space first. Two extractions, both real seams rather than lines
moved to satisfy a gate.

`account_binding.py` holds the two validators that confirm an imported resource
is wired to an account the importer's org actually owns. Shared, not
surface-specific: the schedule step and the surface step hold their account to
the same rule, which is exactly why it is a module and not two private methods.

`surface_apply.py` holds the surface step, which was a fifth of the file on its
own. Nothing about it is shared, which is what makes it the clean cut -- the one
thing it did share is the account-binding rule above.

applier.py: 762 -> 532, and it drops out of `oversized_files` entirely.

Recorded in the baseline, because a ratchet only ratchets if the new number is
written down; left alone, someone could grow it back to 773 without failing.
Doing that also picked up improvements four other modules had already made and
never recorded -- `agent_surfaces/api/schemas.py` and `whatsapp/service.py` are
both under the ceiling now, and `identity`, `schedule` and `usage` had each
shrunk. Those are not this change's doing; they are just true.

Deliberately not claimed: broad-catch counts are aggregated per module, so
moving `except Exception` between files inside `pod_bundle` pays nothing. That
debt is still there and still worth narrowing to real exception types --
`existing_resources.py` is the place to start, where five lookups swallow
everything to answer "does this exist", so a connection error currently reads
as "absent" and the applier creates a duplicate.
