+++ lemma-backend/architecture-baseline.json
@@ -5,7 +5,7 @@
     "agent:total": 76,
     "agent_surfaces:count": 106,
     "agent_surfaces:max": 5,
-    "agent_surfaces:total": 122,
+    "agent_surfaces:total": 121,
     "apps:count": 4,
     "apps:max": 1,
     "apps:total": 4,
@@ -30,9 +30,9 @@
     "pod_bundle:count": 36,
     "pod_bundle:max": 2,
     "pod_bundle:total": 41,
-    "schedule:count": 9,
-    "schedule:max": 2,
-    "schedule:total": 10,
+    "schedule:count": 8,
+    "schedule:max": 1,
+    "schedule:total": 8,
     "workflow:count": 10,
     "workflow:max": 2,
     "workflow:total": 12,
@@ -83,8 +83,6 @@
   "module_cycles": [],
   "oversized_files": {
     "app/modules/agent_surfaces/api/controllers/surface_controller.py": 627,
-    "app/modules/agent_surfaces/api/schemas.py": 578,
-    "app/modules/agent_surfaces/platforms/whatsapp/service.py": 623,
     "app/modules/apps/services/app_service.py": 628,
     "app/modules/connectors/services/connector_operation_service.py": 665,
     "app/modules/connectors/services/connector_service.py": 1423,
@@ -92,12 +90,11 @@
     "app/modules/datastore/infrastructure/kreuzberg_helper.py": 661,
     "app/modules/datastore/services/files/writer.py": 613,
     "app/modules/function/services/function_service.py": 620,
-    "app/modules/identity/services/organization_service.py": 679,
+    "app/modules/identity/services/organization_service.py": 661,
     "app/modules/pod_bundle/events/handlers.py": 1007,
-    "app/modules/pod_bundle/infrastructure/applier.py": 773,
     "app/modules/pod_bundle/infrastructure/exporter.py": 941,
-    "app/modules/schedule/services/schedule_service.py": 654,
-    "app/modules/usage/services/usage_service.py": 696,
+    "app/modules/schedule/services/schedule_service.py": 652,
+    "app/modules/usage/services/usage_service.py": 691,
     "app/modules/workflow/api/schemas.py": 604
   }
 }
